+++ packages/grafana-ui/src/utils/handleReducedMotion.ts
