+++ packages/grafana-ui/src/utils/index.ts
@@ -19,6 +19,6 @@ export { createLogger } from './logger';
 export { attachDebugger } from './debug';
 export * from './nodeGraph';
 export { fuzzyMatch } from './fuzzy';
-export { handleReduceMotion } from './handleReduceMotion';
+export { handleReducedMotion } from './handleReducedMotion';
 
 export { ReactUtils };
